test(mmr): move storage_adapter tests to tests/ dir and cover production error paths
Move inline tests from storage_adapter.rs to tests/test_storage_adapter.rs
so the mock/stub infrastructure (~250 lines of trait stubs) is excluded by
Codecov's existing `**/tests/**` ignore rule. This removes ~440 uncovered
test lines from the coverage denominator.

Add 3 new tests covering the remaining production error paths:
- with_key_size constructor (lines 46-48)
- read key overflow with U32 (line 55)
- write key overflow with U32 (line 86)

The only remaining uncovered lines (90-95) are the serialize error path
in append(), which is defensive code that cannot be triggered through
the public API since MmrNode::serialize() is infallible.
