+++ grovedb-merkle-mountain-range/src/tests/mod.rs
@@ -2,3 +2,5 @@ mod test_coverage;
 mod test_helper;
 mod test_incremental;
 mod test_mmr;
+#[cfg(feature = "storage")]
+mod test_storage_adapter;
